+++ test/usecases/repo_usecases_test.dart
@@ -109,6 +109,10 @@ class MockRepoRepository implements RepoRepository {
       throw UnimplementedError();
 
   @override
+  Future<Either<Failure, Repository>> createRepo(Map<String, dynamic> body) =>
+      throw UnimplementedError();
+
+  @override
   Future<Either<Failure, Branch>> getBranch(
     String owner,
     String repo,